+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # stoker
 
+## 0.4.198
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/node-client@0.5.69
+    - @stoker-platform/types@0.5.48
+    - @stoker-platform/web-app@0.5.171
+    - @stoker-platform/cli@0.5.103
+    - @stoker-platform/utils@0.5.60
+    - @stoker-platform/web-client@0.5.70
+
 ## 0.4.197
 
 ### Patch Changes