+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stoker-platform/utils
 
+## 0.5.60
+
+### Patch Changes
+
+- Updated dependencies
+    - @stoker-platform/types@0.5.48
+
 ## 0.5.59
 
 ### Patch Changes